lambda_build_temp/requirements-lambda.txt
boto3>=1.34.0
python-dotenv>=1.0.0
pydantic>=2.6.1
pydantic-settings>=2.2.1
numpy<2.0,>=1.19.0
pymilvus>=2.3.0
grpcio>=1.48.0
protobuf>=3.20.0
ujson>=5.0.0
pandas<2.0.0
